+++ CHANGELOG.md
@@ -5,9 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file follows the convention described at
 [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## Unreleased
+## [1.1.0] - 2025-08-12
 ### Added
 - Implementation of `IndexMut` trait and `get_mut()` function.
+- Functions `push_within_capacity()`, `pop_if()`, `capacity()`, and `swap_remove()`
 
 ## [1.0.0] - 2025-08-11
 ### Changed